use ResolveType everywhere and move it into execution package move UnboxPossibleOptional into execution
